refactor: Tidy constants, justify suppressions, reorder funcs, split tests
- constants.py: blank lines before comment-attached constants.
- Remove unneeded core_models __all__; rename pty fds master/slave -> parent/child.
- Justify or eliminate every noqa/ty:ignore (fix underlying issue where easy:
  drop dead F841 vars, convert E731 lambdas to defs, fix several ty:ignores);
  no bare suppressions remain.
- Order functions high-level-first (main/public entry first) across modules.
- Split the leaderboards-package tests out of test_process_evaluation_queue.py
  into tests/leaderboards/test_evaluation_common.py and test_queue_hf_cache.py.
