+++ src/leaderboards/link_generation.py
@@ -165,101 +165,6 @@ def ask_user_to_remove_model(model_id: str) -> bool:
 
 
 @cache
-def _load_model_url_decisions() -> dict[str, bool]:
-    """Load cached model URL decisions (remove or keep).
-
-    Returns:
-        A dict mapping model IDs to whether they should be removed (True)
-        or kept without a URL (False). Returns an empty dict if the cache
-        file does not exist.
-    """
-    if not MODELS_WITHOUT_URLS_CACHE.exists():
-        return {}
-    with MODELS_WITHOUT_URLS_CACHE.open("r") as f:
-        data = safe_load(f) or {}
-    # Backwards compatibility: old format was a list of model IDs to keep
-    if isinstance(data, list):
-        return {model_id: False for model_id in data}
-    return data
-
-
-def _load_model_url_decision(model_id: str) -> bool | None:
-    """Load a cached decision for a specific model.
-
-    Args:
-        model_id:
-            The model ID to look up.
-
-    Returns:
-        True if the model should be removed, False if it should be kept
-        without a URL, or None if no cached decision exists.
-    """
-    decisions = _load_model_url_decisions()
-    return decisions.get(model_id)
-
-
-def _remember_model_url_decision(model_id: str, remove: bool) -> None:
-    """Persist a model URL decision to the cache.
-
-    Args:
-        model_id:
-            The model ID.
-        remove:
-            True if the model should be removed, False if it should be
-            kept without a URL.
-    """
-    decisions = _load_model_url_decisions()
-    if model_id in decisions:
-        return
-    decisions[model_id] = remove
-    with MODELS_WITHOUT_URLS_CACHE.open("w") as f:
-        safe_dump(dict(sorted(decisions.items())), f)
-    _load_model_url_decisions.cache_clear()
-    generate_anchor_tag.cache_clear()
-
-
-def _check_model_exists_with_retry(model_id: str, hf_api: HfApi) -> None:
-    """Check if a model exists on the Hugging Face Hub with retry logic.
-
-    Retries only on connection-related errors (not repository errors).
-
-    Args:
-        model_id:
-            The Hugging Face model ID.
-        hf_api:
-            The Hugging Face API client.
-
-    Raises:
-        RepositoryNotFoundError:
-            If the repository does not exist (not retried).
-        GatedRepoError:
-            If the repository is gated (not retried).
-        HFValidationError:
-            If the model ID is invalid (not retried).
-        httpx.RemoteProtocolError:
-            If the server disconnects without sending a response (retried).
-        ConnectionError:
-            If there is a network connection error (retried).
-    """
-    max_attempts = 3
-    for attempt in range(max_attempts):
-        try:
-            hf_api.model_info(repo_id=model_id)
-            return
-        except (httpx.RemoteProtocolError, ConnectionError) as e:
-            if attempt == max_attempts - 1:
-                raise  # Re-raise on last attempt
-            wait_time = 2**attempt  # Exponential backoff: 1s, 2s, 4s
-            logger.warning(
-                f"Connection error checking {model_id}: {e}. "
-                f"Retrying in {wait_time}s..."
-            )
-            time.sleep(wait_time)
-        except (RepositoryNotFoundError, GatedRepoError, HFValidationError):
-            raise  # Don't retry these errors
-
-
-@cache
 def generate_hf_hub_url(model_id: str) -> str | None:
     """Generate a model URL for a model hosted on the Hugging Face Hub.
 
@@ -416,3 +321,98 @@ def generate_ordbogen_url(model_id: str) -> str | None:
         return None
     model_id = model_id.replace("ordbogen/", "")
     return f"https://www.ordbogen.ai/docs/models/{model_id}"
+
+
+def _check_model_exists_with_retry(model_id: str, hf_api: HfApi) -> None:
+    """Check if a model exists on the Hugging Face Hub with retry logic.
+
+    Retries only on connection-related errors (not repository errors).
+
+    Args:
+        model_id:
+            The Hugging Face model ID.
+        hf_api:
+            The Hugging Face API client.
+
+    Raises:
+        RepositoryNotFoundError:
+            If the repository does not exist (not retried).
+        GatedRepoError:
+            If the repository is gated (not retried).
+        HFValidationError:
+            If the model ID is invalid (not retried).
+        httpx.RemoteProtocolError:
+            If the server disconnects without sending a response (retried).
+        ConnectionError:
+            If there is a network connection error (retried).
+    """
+    max_attempts = 3
+    for attempt in range(max_attempts):
+        try:
+            hf_api.model_info(repo_id=model_id)
+            return
+        except (httpx.RemoteProtocolError, ConnectionError) as e:
+            if attempt == max_attempts - 1:
+                raise  # Re-raise on last attempt
+            wait_time = 2**attempt  # Exponential backoff: 1s, 2s, 4s
+            logger.warning(
+                f"Connection error checking {model_id}: {e}. "
+                f"Retrying in {wait_time}s..."
+            )
+            time.sleep(wait_time)
+        except (RepositoryNotFoundError, GatedRepoError, HFValidationError):
+            raise  # Don't retry these errors
+
+
+def _remember_model_url_decision(model_id: str, remove: bool) -> None:
+    """Persist a model URL decision to the cache.
+
+    Args:
+        model_id:
+            The model ID.
+        remove:
+            True if the model should be removed, False if it should be
+            kept without a URL.
+    """
+    decisions = _load_model_url_decisions()
+    if model_id in decisions:
+        return
+    decisions[model_id] = remove
+    with MODELS_WITHOUT_URLS_CACHE.open("w") as f:
+        safe_dump(dict(sorted(decisions.items())), f)
+    _load_model_url_decisions.cache_clear()
+    generate_anchor_tag.cache_clear()
+
+
+def _load_model_url_decision(model_id: str) -> bool | None:
+    """Load a cached decision for a specific model.
+
+    Args:
+        model_id:
+            The model ID to look up.
+
+    Returns:
+        True if the model should be removed, False if it should be kept
+        without a URL, or None if no cached decision exists.
+    """
+    decisions = _load_model_url_decisions()
+    return decisions.get(model_id)
+
+
+@cache
+def _load_model_url_decisions() -> dict[str, bool]:
+    """Load cached model URL decisions (remove or keep).
+
+    Returns:
+        A dict mapping model IDs to whether they should be removed (True)
+        or kept without a URL (False). Returns an empty dict if the cache
+        file does not exist.
+    """
+    if not MODELS_WITHOUT_URLS_CACHE.exists():
+        return {}
+    with MODELS_WITHOUT_URLS_CACHE.open("r") as f:
+        data = safe_load(f) or {}
+    # Backwards compatibility: old format was a list of model IDs to keep
+    if isinstance(data, list):
+        return {model_id: False for model_id in data}
+    return data
