+++ src/scripts/dataset_creation/create_allocine.py
@@ -11,7 +11,10 @@
 """Create the Allocine-mini sentiment dataset and upload it to the HF Hub."""
 
 import pandas as pd
-from constants import MAX_NUM_CHARS_IN_DOCUMENT, MIN_NUM_CHARS_IN_DOCUMENT  # noqa
+
+# These constants are used only inside pandas .query() strings, so the linter
+# cannot see the use.
+from constants import MAX_NUM_CHARS_IN_DOCUMENT, MIN_NUM_CHARS_IN_DOCUMENT  # noqa: F401
 from datasets import Dataset, DatasetDict, Split, load_dataset
 from huggingface_hub import HfApi
