+++ src/scripts/dataset_creation/create_hotter_and_colder_sentiment.py
@@ -25,7 +25,10 @@ import joblib
 import pandas as pd
 import requests as rq
 from bs4 import BeautifulSoup, NavigableString, Tag
-from constants import MAX_NUM_CHARS_IN_DOCUMENT, MIN_NUM_CHARS_IN_DOCUMENT  # noqa
+
+# These constants are used only inside pandas .query() strings, so the linter
+# cannot see the use.
+from constants import MAX_NUM_CHARS_IN_DOCUMENT, MIN_NUM_CHARS_IN_DOCUMENT  # noqa: F401
 from datasets import Dataset, DatasetDict, Split
 from huggingface_hub import HfApi
 from tqdm.auto import tqdm
