+++ src/scripts/dataset_creation/create_ice_linguistic.py
@@ -14,7 +14,10 @@ from ast import literal_eval
 
 import pandas as pd
 import requests
-from constants import MAX_NUM_CHARS_IN_DOCUMENT, MIN_NUM_CHARS_IN_DOCUMENT  # noqa
+
+# These constants are used only inside pandas .query() strings, so the linter
+# cannot see the use.
+from constants import MAX_NUM_CHARS_IN_DOCUMENT, MIN_NUM_CHARS_IN_DOCUMENT  # noqa: F401
 from datasets import Dataset, DatasetDict, Split
 from huggingface_hub import HfApi
