+++ tests/test_scripts/test_process_evaluation_queue.py
@@ -1,12 +1,7 @@
 """Tests for the process_evaluation_queue script orchestration."""
 
-from types import SimpleNamespace
-
 import pytest
-from huggingface_hub import ModelInfo
 
-from leaderboards import evaluation_common
-from leaderboards.queue_hf_cache import is_gguf_model
 from src.scripts import process_evaluation_queue
 
 
@@ -251,88 +246,3 @@ def test_process_issue_does_not_special_case_oom_anymore(
     assert len(comments) == 1
     assert "euroeval exited with code 1" in comments[0]
     assert unassigned == [42]
-
-
-@pytest.mark.parametrize(
-    argnames=["dtype", "count", "expected_bytes"],
-    argvalues=[
-        ("INT4", 3, 2),
-        ("custom_8", 9, 9),
-        ("something16_else", 1, 2),
-        ("unknown_dtype", 5, 10),
-    ],
-)
-def test_estimated_model_bytes_dtype_fallback(
-    monkeypatch: pytest.MonkeyPatch, dtype: str, count: int, expected_bytes: int
-) -> None:
-    """Unknown dtypes should use numeric fallback, then BF16 fallback."""
-
-    def fake_get_safetensors_metadata(
-        *, repo_id: str, revision: str | None = None, token: str | None = None
-    ) -> SimpleNamespace:
-        _ = repo_id, revision, token
-        return SimpleNamespace(parameter_count={dtype: count})
-
-    monkeypatch.setattr(
-        evaluation_common, "get_safetensors_metadata", fake_get_safetensors_metadata
-    )
-    bytes_needed = evaluation_common.estimated_model_bytes(model_id="org/model-name")
-    assert bytes_needed == expected_bytes
-
-
-@pytest.mark.parametrize(
-    argnames=["model_id"], argvalues=[("org/model#low@rev",), ("org/model@rev#low",)]
-)
-def test_estimated_model_bytes_handles_model_id_extras(
-    monkeypatch: pytest.MonkeyPatch, model_id: str
-) -> None:
-    """Model id parsing should handle # and @ in either order."""
-    calls: list[dict[str, str | None]] = []
-
-    def fake_get_safetensors_metadata(
-        *, repo_id: str, revision: str | None = None, token: str | None = None
-    ) -> SimpleNamespace:
-        calls.append({"repo_id": repo_id, "revision": revision, "token": token})
-        return SimpleNamespace(parameter_count={"BF16": 1})
-
-    monkeypatch.setattr(
-        evaluation_common, "get_safetensors_metadata", fake_get_safetensors_metadata
-    )
-    assert evaluation_common.estimated_model_bytes(model_id=model_id) == 2
-    assert calls[0]["repo_id"] == "org/model"
-    assert calls[0]["revision"] == "rev"
-
-
-@pytest.mark.parametrize(
-    argnames=["info", "expected"],
-    argvalues=[
-        # The "gguf" tag alone is enough, regardless of file layout.
-        (ModelInfo(id="org/m", tags=["gguf", "qwen"]), True),
-        # Tag matching is case-insensitive.
-        (ModelInfo(id="org/m", tags=["GGUF"]), True),
-        # library_name set to gguf.
-        (ModelInfo(id="org/m", library_name="gguf"), True),
-        # A .gguf file nested in a per-quant subfolder, no tag.
-        (
-            ModelInfo(
-                id="org/m", siblings=[{"rfilename": "Q4_K_M/model-00001-of-2.gguf"}]
-            ),
-            True,
-        ),
-        # A plain safetensors model is not GGUF.
-        (
-            ModelInfo(
-                id="org/m",
-                tags=["text-generation"],
-                library_name="transformers",
-                siblings=[{"rfilename": "model.safetensors"}],
-            ),
-            False,
-        ),
-        # Missing/None attributes must not raise.
-        (ModelInfo(id="org/m"), False),
-    ],
-)
-def test_is_gguf_model(info: ModelInfo, expected: bool) -> None:
-    """GGUF repos are detected via tag, library_name, or any .gguf sibling."""
-    assert is_gguf_model(info=info) is expected
